+++ src/import/index.ts
@@ -1,3 +1,3 @@
-export * from './Importer';
-export * from './ImportUtils';
 export * from './parser';
+export * from './ImportUtils';
+export * from './Importer';
