+++ src/index.ts
@@ -1,3 +1,2 @@
-export * from './import';
 export * from './valuetype';
-export {default} from './phovea';
+export * from './import';
