+++ src/phovea.ts
@@ -36,7 +36,7 @@ export default function (registry: IRegistry) {
   });
 
   registry.push('importer_value_type', 'idType', function () {
-    return import('./valuetype/valuetype_idtype').then((v) => v.IDTypeUtils);
+    return import('./valuetype/idtypes').then((v) => v.IDTypeUtils);
   }, {
     'factory': 'idType',
     'name': 'IDType',
@@ -49,6 +49,5 @@ export default function (registry: IRegistry) {
   }, <ILocaleEPDesc>{
     ns: 'phovea',
   });
-
 
 }
