+++ src/valuetype/index.ts
@@ -1,2 +1,2 @@
-export * from './valuetype_idtype';
 export * from './valuetypes';
+export * from './idtypes';
